refactor: circuit mod and errors
refactor: circuit module
