+++ src/da/mod.rs
@@ -1,9 +1,10 @@
 use crate::{
+    circuits::{Bls12Proof, VerifyingKey},
     common::Operation,
-    error::{DAResult, GeneralError, PrismResult},
+    error::GeneralError,
     utils::SignedContent,
-    zk_snark::{Bls12Proof, VerifyingKey},
 };
+use anyhow::Result;
 use async_trait::async_trait;
 use borsh::{BorshDeserialize, BorshSerialize};
 use ed25519::Signature;
@@ -25,7 +26,7 @@ pub struct FinalizedEpoch {
 }
 
 impl SignedContent for FinalizedEpoch {
-    fn get_signature(&self) -> PrismResult<Signature> {
+    fn get_signature(&self) -> Result<Signature> {
         match &self.signature {
             Some(signature) => Signature::from_str(signature)
                 .map_err(|e| GeneralError::ParsingError(format!("signature: {}", e)).into()),
@@ -33,13 +34,13 @@ impl SignedContent for FinalizedEpoch {
         }
     }
 
-    fn get_plaintext(&self) -> PrismResult<Vec<u8>> {
+    fn get_plaintext(&self) -> Result<Vec<u8>> {
         let mut copy = self.clone();
         copy.signature = None;
         borsh::to_vec(&copy).map_err(|e| GeneralError::EncodingError(e.to_string()).into())
     }
 
-    fn get_public_key(&self) -> PrismResult<String> {
+    fn get_public_key(&self) -> Result<String> {
         //TODO(@distractedm1nd): the below comment isn't good enough of an argument to not return the public key, it should be fixed
 
         // for epoch json the public key to verify is the one from the sequencer which should be already be public and known from every light client
@@ -50,11 +51,11 @@ impl SignedContent for FinalizedEpoch {
 
 #[async_trait]
 pub trait DataAvailabilityLayer: Send + Sync {
-    async fn get_latest_height(&self) -> DAResult<u64>;
-    async fn initialize_sync_target(&self) -> DAResult<u64>;
-    async fn get_snarks(&self, height: u64) -> DAResult<Vec<FinalizedEpoch>>;
-    async fn submit_snarks(&self, epoch: Vec<FinalizedEpoch>) -> DAResult<u64>;
-    async fn get_operations(&self, height: u64) -> DAResult<Vec<Operation>>;
-    async fn submit_operations(&self, operations: Vec<Operation>) -> DAResult<u64>;
-    async fn start(&self) -> DAResult<()>;
+    async fn get_latest_height(&self) -> Result<u64>;
+    async fn initialize_sync_target(&self) -> Result<u64>;
+    async fn get_snarks(&self, height: u64) -> Result<Vec<FinalizedEpoch>>;
+    async fn submit_snarks(&self, epoch: Vec<FinalizedEpoch>) -> Result<u64>;
+    async fn get_operations(&self, height: u64) -> Result<Vec<Operation>>;
+    async fn submit_operations(&self, operations: Vec<Operation>) -> Result<u64>;
+    async fn start(&self) -> Result<()>;
 }
