+++ src/node_types/mod.rs
@@ -1,4 +1,4 @@
-use crate::error::PrismResult;
+use anyhow::Result;
 use async_trait::async_trait;
 use std::{self, sync::Arc};
 
@@ -7,6 +7,6 @@ pub mod sequencer;
 
 #[async_trait]
 pub trait NodeType {
-    async fn start(self: Arc<Self>) -> PrismResult<()>;
+    async fn start(self: Arc<Self>) -> Result<()>;
     // async fn stop(&self) -> Result<(), String>;
 }
